refactor small; add other imports
