+++ sklearnex/cluster/tests/test_dbscan.py
@@ -18,12 +18,14 @@
 import numpy as np
 from numpy.testing import assert_allclose
 
+
 def test_sklearnex_import():
     from sklearnex.cluster import DBSCAN
     X = np.array([[1, 2], [2, 2], [2, 3],
                   [8, 7], [8, 8], [25, 80]])
     dbscan = DBSCAN(eps=3, min_samples=2).fit(X)
+    assert 'daal4py' in dbscan.__module__
 
     result = dbscan.labels_
-    expected = np.array([ 0, 0, 0, 1, 1, -1], dtype=np.int32)
+    expected = np.array([0, 0, 0, 1, 1, -1], dtype=np.int32)
     assert_allclose(expected, result)
