+++ sklearnex/ensemble/tests/test_random_forest.py
@@ -17,11 +17,23 @@
 
 import numpy as np
 from numpy.testing import assert_allclose
-from sklearn.datasets import make_regression
+from sklearn.datasets import make_classification, make_regression
 
-def test_sklearnex_import():
+
+def test_sklearnex_import_rf_classifier():
+    from sklearnex.ensemble import RandomForestClassifier
+    X, y = make_classification(n_samples=1000, n_features=4,
+                               n_informative=2, n_redundant=0,
+                               random_state=0, shuffle=False)
+    rf = RandomForestClassifier(max_depth=2, random_state=0).fit(X, y)
+    assert 'daal4py' in rf.__module__
+    assert_allclose([1], rf.predict([[0, 0, 0, 0]]))
+
+
+def test_sklearnex_import_rf_regression():
     from sklearnex.ensemble import RandomForestRegressor
     X, y = make_regression(n_features=4, n_informative=2,
                            random_state=0, shuffle=False)
     rf = RandomForestRegressor(max_depth=2, random_state=0).fit(X, y)
+    assert 'daal4py' in rf.__module__
     assert_allclose([-6.66], rf.predict([[0, 0, 0, 0]]), atol=1e-2)
