+++ sklearnex/manifold/tests/test_tsne.py
@@ -17,12 +17,10 @@
 
 import numpy as np
 from numpy.testing import assert_allclose
-from sklearn.datasets import load_iris
 
 
 def test_sklearnex_import():
-    from sklearnex.linear_model import LogisticRegression
-    X, y = load_iris(return_X_y=True)
-    logreg = LogisticRegression(random_state=0, max_iter=200).fit(X, y)
-    assert 'daal4py' in logreg.__module__
-    assert_allclose(logreg.score(X, y), 0.9733, atol=1e-3)
+    from sklearnex.manifold import TSNE
+    X = np.array([[0, 0, 0], [0, 1, 1], [1, 0, 1], [1, 1, 1]])
+    tsne = TSNE(n_components=2).fit(X)
+    assert 'daal4py' in tsne.__module__
